third_party/blink/web_tests/fast/dom/custom/lifecycle-created-creation-api-expected.txt

CONSOLE WARNING: line 13: document.registerElement is deprecated and will be removed in M80, around February 2020. Please use window.customElements.define instead. See https://www.chromestatus.com/features/4642138092470272 and https://developers.google.com/web/updates/2019/07/web-components-time-to-upgrade for more details.
This test ensures that the lifecycle callbacks of API-originated elements are visible in following script block.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS document.createElement('x-foo').callbacksCalled is true
PASS document.createElement('div', 'x-bar').callbacksCalled is true
PASS document.createElementNS('http://www.w3.org/1999/xhtml', 'x-foo').callbacksCalled is true
PASS document.createElementNS('http://www.w3.org/1999/xhtml', 'div', 'x-bar').callbacksCalled is true
PASS foo.cloneNode().callbacksCalled is true
PASS foo.cloneNode(true).firstChild.callbacksCalled is true
PASS bar.cloneNode().callbacksCalled is true
PASS bar.cloneNode(true).firstChild.callbacksCalled is true
PASS foreignFoo.callbacksCalled is true
PASS foreignFoo.firstChild.callbacksCalled is true
PASS importedFoo.callbacksCalled is true
PASS importedFoo.firstChild.callbacksCalled is true
PASS window.callbacksCalled is ['div', 'x-foo']
PASS importedBar.callbacksCalled is true
PASS importedBar.firstChild.callbacksCalled is true
PASS window.callbacksCalled is ['DIV', 'X-FOO']
PASS window.callbacksCalled is ['X-FOO']
PASS window.callbacksCalled is ['X-FOO']
PASS successfullyParsed is true

TEST COMPLETE
